backend/pyproject.toml
[tool.poetry]
name = "tasks-api"
version = "0.1.0"
description = ""
authors = ["Felipe Carvalho <felipe.carvalho07@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.5"
djangorestframework = "^3.16.1"
gunicorn = "^23.0.0"
pyjwt = "^2.10.1"
dotenv = "^0.9.9"
dj-database-url = "^3.0.1"
psycopg2-binary = "^2.9.10"

[tool.poetry.group.dev.dependencies]
isort = "^6.0.1"
flake8 = "^7.3.0"
black = "^25.1.0"
pre-commit = "^4.3.0"

[tool.black]
line-length = 120
target-version = ["py312"]
skip-string-normalization = true

[tool.isort]
profile = "black"
line_length = 120
known_first_party = ["server"]
sections = "FUTURE,STDLIB,DJANGO,THIRDPARTY,FIRSTPARTY,LOCALFOLDER"
known_django = "django"
known_local_folder= ["core", "user_profiles"]
default_section = "THIRDPARTY"
multi_line_output = 2
include_trailing_comma = true
use_parentheses = true

[tool.flake8]
max-line-length = 120
extend-ignore = [
    "E203",
    "E501",
    "W503",
]
exclude = [
    ".git",
    "__pycache__",
    "build",
    "dist",
    "migrations",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
